added audio value modulation and cool icons
