--- .idea/.idea.LongTermGame/.idea/codeStyles/codeStyleConfig.xml
@@ -1,5 +1,0 @@
-﻿<component name="ProjectCodeStyleConfiguration">
-  <state>
-    <option name="PREFERRED_PROJECT_CODE_STYLE" value="Default" />
-  </state>
-</component>
