--- .idea/.idea.LongTermGame/.idea/contentModel.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ContentModelStore">
-    <e p="C:\Users\Daniel\AppData\Local\JetBrains\Rider2020.2\extResources" t="IncludeRecursive" />
-    <e p="C:\Users\Daniel\AppData\Local\JetBrains\Rider2020.2\resharper-host\Local\Transient\Rider\v202\SolutionCaches\_LongTermGame.-2003281793.00" t="ExcludeRecursive" />
-    <e p="C:\Users\Daniel\Documents\Programming\Projects\LongTermGame" t="IncludeRecursive">
-      <e p="Assembly-CSharp-Editor.csproj" t="IncludeRecursive" />
-      <e p="Assembly-CSharp.csproj" t="IncludeRecursive" />
-      <e p="Assets" t="Include">
-        <e p="arrowTest.cs" t="Include" />
-        <e p="PlayerManager.cs" t="Include" />
-        <e p="Scripts" t="Include">
-          <e p="ArrowReflector.cs" t="Include" />
-          <e p="ArrowScript.cs" t="Include" />
-          <e p="ArrowTriggeredButton.cs" t="Include" />
-          <e p="Audio Scripts" t="Include">
-            <e p="AudioManager.cs" t="Include" />
-            <e p="Sound.cs" t="Include" />
-            <e p="TestSounds.cs" t="Include" />
-          </e>
-          <e p="ButtonManager.cs" t="Include" />
-          <e p="CameraFX.cs" t="Include" />
-          <e p="GrapplingArrow.cs" t="Include" />
-          <e p="Looking.cs" t="Include" />
-          <e p="PlayerMovement.cs" t="Include" />
-          <e p="PlayerShooting.cs" t="Include" />
-          <e p="SimpleCameraController.cs" t="Include" />
-          <e p="SlidingDoor.cs" t="Include" />
-          <e p="TestPlayer.cs" t="Include" />
-        </e>
-        <e p="TutorialInfo" t="Include">
-          <e p="Scripts" t="Include">
-            <e p="Editor" t="Include">
-              <e p="ReadmeEditor.cs" t="Include" />
-            </e>
-            <e p="Readme.cs" t="Include" />
-          </e>
-        </e>
-      </e>
-      <e p="LongTermGame.sln" t="IncludeFlat" />
-      <e p="obj" t="ExcludeRecursive" />
-      <e p="Packages" t="ExcludeRecursive" />
-      <e p="Temp\bin" t="ExcludeRecursive" />
-    </e>
-  </component>
-</project>
