--- .idea/.idea.LongTermGame/.idea/indexLayout.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ContentModelUserStore">
-    <attachedFolders />
-    <explicitIncludes>
-      <Path>Library/PackageCache/com.unity.collab-proxy@1.2.16</Path>
-      <Path>Library/PackageCache/com.unity.ext.nunit@1.0.0</Path>
-      <Path>Library/PackageCache/com.unity.ide.rider@1.1.4</Path>
-      <Path>Library/PackageCache/com.unity.ide.vscode@1.2.1</Path>
-      <Path>Library/PackageCache/com.unity.render-pipelines.core@7.3.1</Path>
-      <Path>Library/PackageCache/com.unity.render-pipelines.universal@7.3.1</Path>
-      <Path>Library/PackageCache/com.unity.shadergraph@7.3.1</Path>
-      <Path>Library/PackageCache/com.unity.test-framework@1.1.16</Path>
-      <Path>Library/PackageCache/com.unity.textmeshpro@2.1.1</Path>
-      <Path>Library/PackageCache/com.unity.timeline@1.2.16</Path>
-      <Path>Packages</Path>
-      <Path>ProjectSettings</Path>
-    </explicitIncludes>
-    <explicitExcludes>
-      <Path>.git</Path>
-      <Path>.idea</Path>
-      <Path>.vs</Path>
-      <Path>.vscode</Path>
-      <Path>GitHub Documentation</Path>
-      <Path>Library</Path>
-      <Path>Temp</Path>
-      <Path>ThePeople</Path>
-      <Path>obj</Path>
-    </explicitExcludes>
-  </component>
-</project>
