--- .idea/.idea.LongTermGame/.idea/modules.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/.idea.LongTermGame/riderModule.iml" filepath="$PROJECT_DIR$/.idea/.idea.LongTermGame/riderModule.iml" />
-    </modules>
-  </component>
-</project>
