+++ Assets/Audio/Slight Random.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df2d2bc52b5c7814b9909d02430121ac
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
