+++ Assets/Scripts/Audio/AudioManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1a7f6f2cf90269488aa2e40e66e6059
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 49a969f4aecce2d43895c132c74ab207, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
