+++ Assets/Scripts/Audio/Modulator.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: ca2e730c346ae974a8df7cead93c6e50
+guid: 9aefe1003c5c42ee90e3c95402b50d2d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {instanceID: 0}
+  icon: {fileID: 2800000, guid: 8e93b0f53210e9948a5f95f858a0e303, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
