+++ Assets/Scripts/Audio/Modulators/Random.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: d1a7f6f2cf90269488aa2e40e66e6059
+guid: e6f30b3125ad48a48fb9bd7edaedd6a2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {instanceID: 0}
+  icon: {fileID: 2800000, guid: 8e93b0f53210e9948a5f95f858a0e303, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
