+++ Assets/Scripts/Audio/Sound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca2e730c346ae974a8df7cead93c6e50
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 413ca9c70b8916e43a4ec6d5f75052c8, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
